pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yaml-workflow-engine"
version = "0.1.0"
description = "Engine para execução de workflows declarativos em YAML com handlers registráveis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{name = "Eric Jones Silva", email = "ericjonsilva@outlook.com"}]
keywords = ["yaml", "workflow", "automação", "engine", "rpa", "botcity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
http = ["requests>=2.28"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
